feat: move transport module into core module (#2)
