+++ src/core.rs
@@ -1,6 +1,4 @@
 // Copyright (C) 2026 ReteLabs LLC.
 // Licensed under Apache-2.0 or MIT at your option.
 
-pub(crate) mod actor;
-pub mod core;
-pub mod logging;
+pub mod transport;
