+++ src/core/transport/resolver.rs
